[docs-infra] Replace mapApiPageTranslations with singular mapApiPageTranslation
Collapse the helper to operate on a single raw translation object. The
caller no longer wraps in a require.context shim or a per-language map;
it just passes the imported English JSON. The function mutates and
returns the translation with `componentDescription` rendered to HTML
and `componentDescriptionToc` populated.

Breaking change for downstream consumers (mui-x). Coordinate the
@mui/internal-core-docs publish with PR 3+4 once mui-x has migrated.
